% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{estat_api}
\alias{estat_api}
\title{e-Stat API}
\usage{
estat_api(path, appId, ...)
}
\arguments{
\item{path}{API endpoint.}

\item{appId}{Application ID.}

\item{...}{Other parameters.}
}
\description{
Get Statistical Something From e-Stat API
}
