% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataCatalog.R
\name{estat_getDataCatalog}
\alias{estat_getDataCatalog}
\title{getDataCatalog API}
\usage{
estat_getDataCatalog(appId, ...)
}
\arguments{
\item{appId}{Application ID}

\item{...}{Other parameters.}
}
\description{
Get information about the statistical dataset files and databases via e-Stat API.
}
\seealso{
\url{http://www.e-stat.go.jp/api/e-stat-manual/#api_2_6}
\url{http://www.e-stat.go.jp/api/e-stat-manual/#api_3_7}
}

