% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaInfo.R
\name{estat_getMetaInfo}
\alias{estat_getMetaInfo}
\title{getMetaInfo API}
\usage{
estat_getMetaInfo(appId, statsDataId, lang = c("J", "E"), ...)
}
\arguments{
\item{appId}{Application ID.}

\item{statsDataId}{ID of the statistical dataset.}

\item{lang}{Language of data. \code{"J"}(Japanese) or \code{"E"}(English).}

\item{...}{Other parameters.}
}
\description{
Get meta-information about the statistical dataset via e-Stat API.
}
\examples{
\dontrun{
estat_getMetaInfo(appId = "XXXX", statsDataId = "0003065345")
}

}
\seealso{
\url{http://www.e-stat.go.jp/api/e-stat-manual3-0/#api_2_2}
\url{http://www.e-stat.go.jp/api/e-stat-manual3-0/#api_3_3}
}
