% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estatapi.R
\docType{package}
\name{estatapi}
\alias{estatapi}
\alias{estatapi-package}
\title{R interface to e-Stat API}
\description{
R interface to e-Stat API
}
\seealso{
\url{http://www.e-stat.go.jp/SG1/estat/eStatTopPortalE.do}
}

