% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataCatalog.R
\name{estat_getDataCatalog}
\alias{estat_getDataCatalog}
\title{getDataCatalog API}
\usage{
estat_getDataCatalog(appId, use_label = TRUE, ...)
}
\arguments{
\item{appId}{Application ID}

\item{use_label}{Whether to take the human-redable label value or the code value when flattening a field containing both.
(default: \code{TRUE})}

\item{...}{Other parameters.}
}
\description{
Get information about the statistical dataset files and databases via e-Stat API.
}
\section{Other parameters}{

For every detailed information, please visit the URL in See Also.
\itemize{
 \item \code{surveyYears}:
   Year and month when the survey was conducted. The format is either \code{YYYY}, \code{YYYYMM}, or \code{YYYYMM-YYYYMM}
 \item \code{openYears}:
   Year and month when the survey result was opened. The format is the same as \code{surveyYears}
 \item \code{statsField}:
   Field of statistics. The format is either two digits (large classification) or
   four digits (small classification). For the detail of the classification, see
   \url{http://www.soumu.go.jp/toukei_toukatsu/index/seido/sangyo/26index.htm}
 \item \code{statsCode}:
    Code assigned for each statistical agency and statistics. The format can be
    five digits (agency), and eight digits (statistics). For the detail, see
    \url{http://www.stat.go.jp/info/guide/public/code/code.htm}.
 \item \code{searchWord}:
    Keyword for searching. You can use \code{OR} and \code{AND}. (e.g.: \code{apple AND orrange}).
 \item \code{dataType}:
    Type of data. \code{XLS}: Excel file, \code{CSV}: CSV file, \code{PDF}: PDF file, \code{DB}: Database.
 \item \code{catalogId}:
    Catalog ID.
 \item \code{resourceId}:
    Catalog resource ID.
 \item \code{startPosition}:
   integer. The the first record to get.
 \item \code{limit}:
   integer. Max number of records to get.
 \item \code{updatedDate}:
   Last updated date. The format is either \code{YYYY}, \code{YYYYMM}, \code{YYYYMMDD}, \code{YYYYMMDD-YYYYMMDD}
}
}
\examples{
\dontrun{
estat_getDataCatalog(
  appId = "XXXX",
  searchWord = "CD",
  dataType = "CSV",
  limit = 3
)
}
}
\seealso{
\url{http://www.e-stat.go.jp/api/e-stat-manual/#api_2_6}
\url{http://www.e-stat.go.jp/api/e-stat-manual/#api_3_7}
}

