% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estatapi-package.R
\docType{package}
\name{estatapi}
\alias{estatapi}
\alias{estatapi-package}
\title{R Interface to e-Stat API}
\description{
Provides an interface to e-Stat API, the one-stop service for official statistics of the Japanese government.
}
\seealso{
\url{http://www.e-stat.go.jp/SG1/estat/eStatTopPortalE.do}
}
\author{
\strong{Maintainer}: Hiroaki Yutani \email{yutani.ini@gmail.com} (\href{https://orcid.org/0000-0002-3385-7233}{ORCID})

}
